Codegen for shuffled hash join
